feat(scintillator/animation): generate keyframes from xml
I found a library called "keytime" that does exactly what I need:
given a set of keyframes and the time, return the state of the animation.

An animation can work like this:

```jade
sprite(...)
  animation
    keyframe(t='0' x='-100')
    keyframe(t='1' x='0')
  animation(on='exitEvent')
    keyframe(t='0' x='0')
    keyframe(t='1' x='1000')
```

Here, there are two animations. The first one is the default animation, and the
second one is the "on exit" animation.

Which animation to use is selected based on latest event. An event is the
timestamp in the data passed to `render()` of the specified key.

    { t: 1.5, exitEvent: 1.2 }

This means the exitEvent has happened 0.3 seconds ago, so the exit animation
should be performed.