feat(SamplesLoader): try .ogg before .mp3
.ogg samples generally gives higher quality over file size and lower latency.

However, Apple is refusing to support .ogg audios [1]. Converting audios to high quality .aac/.m4a is complicated and have different commands on different platforms [2]. Therefore, we use .mp3 for now.

One flaw when using MP3 format is that some silence will be added in front of the audio data both when encoding and decoding, resulting in ~1056 samples delay when playing [3]. This is about 23 milliseconds, which can turn a METICULOUS into a PERFECT. The MP3 audio format should not be used with music games! It is in our plan in the future to escape from .MP3 and go to .M4A/.AAC instead!

[1]: http://en.wikipedia.org/wiki/HTML5_Audio
[2]: http://wiki.hydrogenaud.io/index.php?title=AAC_encoders
[3]: http://lame.sourceforge.net/tech-FAQ.txt