docs(~): link docs to source code
Instead of embedding documentation inside source code, we reference the source
code from the documentation. This makes it easy for other people to edit the
documentation.

We use the //#region syntax borrowed from Visual Studio to mark source code
region, and then use the `.. src::` directive to link to the source code.

The source code under the specified region is then hash-checked and a red
warning will be generated when the source is out of sync with the docs.