:green_heart: Do not run Danger on forked PRs (yet).
This is due to the fact that CircleCI Environment Variables are not made
available to forked PR builds, since these environment variables contain
sensitive information.

That means for forked PRs, DANGER_GITHUB_API_TOKEN will be unset, causing Danger
to crash.

However, we can't just expose the environment variables for forked PR builds, as
that would essentially make the secret keys public.

To resolve this completely, we need to figure out a way to allow Danger to make
GitHub comments WITHOUT putting the GitHub Personal Access Token of akibot in
the public, but that is an issue for another time.

For now, what we can do is just disable Danger for forked PR.